bot/requirements.txt
discord.py>=2.3.2
google-generativeai>=0.3.2
python-dotenv>=1.0.0
aiohttp>=3.9.1
fastapi>=0.104.1
uvicorn>=0.24.0
aiosqlite>=0.19.0
asyncpg>=0.27.0,<0.30.0
pydantic>=2.5.0
websockets>=12.0
wavelink>=3.2.0
PyNaCl>=1.5.0
python-socketio>=5.10.0
colorama>=0.4.6
youtube-search-python>=1.6.6
yt-dlp>=2023.12.30
supabase>=2.0.0
psutil>=5.9.0
lyricsgenius>=3.0.1